requirements.txt
requests
discord.py
psutil
Pillow
pyttsx3
pyautogui
socket
uuid
tempfile
pyperclip
datetime